NBAProject2/requirements.txt
Django>=5.1,<6.0
SPARQLWrapper>=1.8.5
requests
rdflib
python-dotenv